Added StyledContainer.dart, SoundContainer.dart
