+++ audiopoli_dashboard/lib/SoundContainer.dart
@@ -1,0 +1,15 @@
+import 'package:flutter/cupertino.dart';
+import './StyledContainer.dart';
+
+class SoundContainer extends StatelessWidget {
+  const SoundContainer({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return StyledContainer(
+      widget: Container(
+        child: Text("SoundContainer"),
+      ),
+    );
+  }
+}
